third_party/WebKit/LayoutTests/webaudio/audiobuffersource-late-start-expected.txt

Test the late call of start(0) of BufferSource.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS The output contains all the expected values in the correct order: [0,1].
PASS The index of value change is equal to 5504.
PASS The rendered buffer contains non-zero values after the first sample.
PASS successfullyParsed is true

TEST COMPLETE
